[Field][Fieldset] Refactor to useRenderElement
