+++ packages/react/src/field/control/FieldControl.tsx
@@ -1,11 +1,15 @@
 'use client';
 import * as React from 'react';
-import { useComponentRenderer } from '../../utils/useComponentRenderer';
-import { useFieldControl } from './useFieldControl';
 import { FieldRoot } from '../root/FieldRoot';
 import { useFieldRootContext } from '../root/FieldRootContext';
 import { fieldValidityMapping } from '../utils/constants';
 import { BaseUIComponentProps } from '../../utils/types';
+import { useRenderElement } from '../../utils/useRenderElement';
+import { useEventCallback } from '../../utils/useEventCallback';
+import { useField } from '../useField';
+import { useControlled, useModernLayoutEffect } from '../../utils';
+import { useBaseUiId } from '../../utils/useBaseUiId';
+import { useFieldControlValidation } from './useFieldControlValidation';
 
 /**
  * The form control to label and validate.
@@ -18,20 +22,20 @@ import { BaseUIComponentProps } from '../../utils/types';
  * Documentation: [Base UI Field](https://base-ui.com/react/components/field)
  */
 export const FieldControl = React.forwardRef(function FieldControl(
-  props: FieldControl.Props,
+  componentProps: FieldControl.Props,
   forwardedRef: React.ForwardedRef<HTMLInputElement>,
 ) {
   const {
     render,
     className,
-    id,
+    id: idProp,
     name: nameProp,
-    value,
+    value: valueProp,
     disabled: disabledProp = false,
     onValueChange,
     defaultValue,
-    ...otherProps
-  } = props;
+    ...elementProps
+  } = componentProps;
 
   const { state: fieldState, name: fieldName, disabled: fieldDisabled } = useFieldRootContext();
 
@@ -39,26 +43,108 @@ export const FieldControl = React.forwardRef(function FieldControl(
   const name = fieldName ?? nameProp;
 
   const state: FieldControl.State = React.useMemo(
-    () => ({ ...fieldState, disabled }),
+    () => ({
+      ...fieldState,
+      disabled,
+    }),
     [fieldState, disabled],
   );
 
-  const { getControlProps } = useFieldControl({
-    id,
-    name,
-    disabled,
-    value,
-    defaultValue,
-    onValueChange,
+  const {
+    setControlId,
+    labelId,
+    setTouched,
+    setDirty,
+    validityData,
+    setFocused,
+    setFilled,
+    validationMode,
+  } = useFieldRootContext();
+
+  const { getValidationProps, getInputValidationProps, commitValidation, inputRef } =
+    useFieldControlValidation();
+
+  const id = useBaseUiId(idProp);
+
+  useModernLayoutEffect(() => {
+    setControlId(id);
+    return () => {
+      setControlId(undefined);
+    };
+  }, [id, setControlId]);
+
+  useModernLayoutEffect(() => {
+    const hasExternalValue = valueProp != null;
+    if (inputRef.current?.value || (hasExternalValue && valueProp !== '')) {
+      setFilled(true);
+    } else if (hasExternalValue && valueProp === '') {
+      setFilled(false);
+    }
+  }, [inputRef, setFilled, valueProp]);
+
+  const [value, setValueUnwrapped] = useControlled({
+    controlled: valueProp,
+    default: defaultValue,
+    name: 'FieldControl',
+    state: 'value',
   });
 
-  const { renderElement } = useComponentRenderer({
-    propGetter: getControlProps,
-    render: render ?? 'input',
+  const setValue = useEventCallback(
+    (nextValue: string | number | readonly string[], event: Event) => {
+      setValueUnwrapped(nextValue);
+      onValueChange?.(nextValue, event);
+    },
+  );
+
+  useField({
+    id,
+    commitValidation,
+    value,
+    getValue: () => inputRef.current?.value,
+    controlRef: inputRef,
+  });
+
+  const renderElement = useRenderElement('input', componentProps, {
     ref: forwardedRef,
-    className,
     state,
-    extraProps: otherProps,
+    props: [
+      {
+        id,
+        disabled,
+        name,
+        ref: inputRef,
+        'aria-labelledby': labelId,
+        value,
+        onChange(event) {
+          if (value != null) {
+            setValue(event.currentTarget.value, event.nativeEvent);
+          }
+
+          setDirty(event.currentTarget.value !== validityData.initialValue);
+          setFilled(event.currentTarget.value !== '');
+        },
+        onFocus() {
+          setFocused(true);
+        },
+        onBlur(event) {
+          setTouched(true);
+          setFocused(false);
+
+          if (validationMode === 'onBlur') {
+            commitValidation(event.currentTarget.value);
+          }
+        },
+        onKeyDown(event) {
+          if (event.currentTarget.tagName === 'INPUT' && event.key === 'Enter') {
+            setTouched(true);
+            commitValidation(event.currentTarget.value);
+          }
+        },
+      },
+      getValidationProps(),
+      getInputValidationProps(),
+      elementProps,
+    ],
     customStyleHookMapping: fieldValidityMapping,
   });
