+++ packages/react/src/field/description/FieldDescription.tsx
@@ -1,11 +1,12 @@
 'use client';
 import * as React from 'react';
-import { useComponentRenderer } from '../../utils/useComponentRenderer';
 import { FieldRoot } from '../root/FieldRoot';
 import { useFieldRootContext } from '../root/FieldRootContext';
-import { useFieldDescription } from './useFieldDescription';
 import { fieldValidityMapping } from '../utils/constants';
 import type { BaseUIComponentProps } from '../../utils/types';
+import { useBaseUiId } from '../../utils/useBaseUiId';
+import { useModernLayoutEffect } from '../../utils/useModernLayoutEffect';
+import { useRenderElement } from '../../utils/useRenderElement';
 
 /**
  * A paragraph with additional information about the field.
@@ -14,22 +15,33 @@ import type { BaseUIComponentProps } from '../../utils/types';
  * Documentation: [Base UI Field](https://base-ui.com/react/components/field)
  */
 export const FieldDescription = React.forwardRef(function FieldDescription(
-  props: FieldDescription.Props,
+  componentProps: FieldDescription.Props,
   forwardedRef: React.ForwardedRef<HTMLParagraphElement>,
 ) {
-  const { render, id, className, ...otherProps } = props;
+  const { render, id: idProp, className, ...elementProps } = componentProps;
 
   const { state } = useFieldRootContext(false);
 
-  const { getDescriptionProps } = useFieldDescription({ id });
+  const id = useBaseUiId(idProp);
 
-  const { renderElement } = useComponentRenderer({
-    propGetter: getDescriptionProps,
-    render: render ?? 'p',
+  const { setMessageIds } = useFieldRootContext();
+
+  useModernLayoutEffect(() => {
+    if (!id) {
+      return undefined;
+    }
+
+    setMessageIds((v) => v.concat(id));
+
+    return () => {
+      setMessageIds((v) => v.filter((item) => item !== id));
+    };
+  }, [id, setMessageIds]);
+
+  const renderElement = useRenderElement('p', componentProps, {
     ref: forwardedRef,
-    className,
     state,
-    extraProps: otherProps,
+    props: [{ id }, elementProps],
     customStyleHookMapping: fieldValidityMapping,
   });
