+++ packages/react/src/field/label/FieldLabel.tsx
@@ -1,13 +1,13 @@
 'use client';
 import * as React from 'react';
-import { useComponentRenderer } from '../../utils/useComponentRenderer';
+import { getTarget } from '@floating-ui/react/utils';
 import { FieldRoot } from '../root/FieldRoot';
 import { useFieldRootContext } from '../root/FieldRootContext';
-import { useFieldLabel } from './useFieldLabel';
 import { fieldValidityMapping } from '../utils/constants';
 import { useBaseUiId } from '../../utils/useBaseUiId';
 import { useModernLayoutEffect } from '../../utils/useModernLayoutEffect';
 import type { BaseUIComponentProps } from '../../utils/types';
+import { useRenderElement } from '../../utils/useRenderElement';
 
 /**
  * An accessible label that is automatically associated with the field control.
@@ -16,12 +16,12 @@ import type { BaseUIComponentProps } from '../../utils/types';
  * Documentation: [Base UI Field](https://base-ui.com/react/components/field)
  */
 export const FieldLabel = React.forwardRef(function FieldLabel(
-  props: FieldLabel.Props,
+  componentProps: FieldLabel.Props,
   forwardedRef: React.ForwardedRef<any>,
 ) {
-  const { render, className, id: idProp, ...otherProps } = props;
+  const { render, className, id: idProp, ...elementProps } = componentProps;
 
-  const { setLabelId, state } = useFieldRootContext(false);
+  const { labelId, setLabelId, state, controlId } = useFieldRootContext(false);
 
   const id = useBaseUiId(idProp);
 
@@ -32,15 +32,27 @@ export const FieldLabel = React.forwardRef(function FieldLabel(
     };
   }, [id, setLabelId]);
 
-  const { getLabelProps } = useFieldLabel();
-
-  const { renderElement } = useComponentRenderer({
-    propGetter: getLabelProps,
-    render: render ?? 'label',
+  const renderElement = useRenderElement('label', componentProps, {
     ref: forwardedRef,
-    className,
     state,
-    extraProps: otherProps,
+    props: [
+      {
+        id: labelId,
+        htmlFor: controlId,
+        onMouseDown(event) {
+          const target = getTarget(event.nativeEvent) as HTMLElement | null;
+          if (target?.closest('button,input,select,textarea')) {
+            return;
+          }
+
+          // Prevent text selection when double clicking label.
+          if (!event.defaultPrevented && event.detail > 1) {
+            event.preventDefault();
+          }
+        },
+      },
+      elementProps,
+    ],
     customStyleHookMapping: fieldValidityMapping,
   });
 
@@ -50,5 +62,5 @@ export const FieldLabel = React.forwardRef(function FieldLabel(
 export namespace FieldLabel {
   export type State = FieldRoot.State;
 
-  export interface Props extends BaseUIComponentProps<'div', State> {}
+  export interface Props extends BaseUIComponentProps<'label', State> {}
 }
